EX-17 [Code][Modify] Part 2
