+++ src/pages/dashboard/Records.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import useDocumentTitle from "@/hooks/document/useDocTitle";
 import SpentHistory from "@/components/dashboard/spentHistory/SpentHistory";
-// import AddItem from "@/components/dashboard/addItem/AddItem";
-import OverViewPage from "@/components/dashboard/OverView";
+import ExpensesTotal from "@/components/dashboard/ExpensesTotal";
 import {FormattedUTCDate } from "@/utils/DateFormat";
 import SpentSummary from "@/components/dashboard/SpentSummary";
 import OverviewDate from "@/hooks/OverviewDate";
@@ -11,19 +10,9 @@ import { Link  } from "react-router-dom";
 const OverView = () => {
   
   const { dateRange,timeRange } = OverviewDate();
-  // const {setModal,setChild, isModal} = useModal();
 
-  
-  useDocumentTitle("Dashboard - Overview | Finance Management");
+  useDocumentTitle("Records | Finance Management");
 
-  // if (loading) {
-  //       return (
-  //         <div className='flex justify-center w-full h-screen py-20 dark:text-white'>
-  //           <span className='dark:hidden'><ClipLoader size={30} /></span>
-  //           <span className='hidden dark:block'><ClipLoader color={'#ffffff'} size={30} /></span>
-  //         </div>
-  //       )
-  //     }
   return (
     <div className="grid gap-4 px-4 lg:py-6 mx-auto transition max-w-7xl lg:grid-cols-3 dark:bg-dark dark:text-white overflow-hidden lg:overflow-clip">
 
@@ -32,7 +21,7 @@ const OverView = () => {
 
         {/* Overview Section */}
         <section className="pt-4 sticky top-0 bg-white dark:bg-dark dark:border-medium-dark lg:hidden">
-          <OverViewPage />
+          <ExpensesTotal />
           <div className={`${dateRange.includes('-') ? 'block' : 'hidden'} lg:hidden mt-2 py-1 text-center border-t border-gray-300 dark:border-light-dark`}>
               <p className='text-sm text-gray-400'>{dateRange}</p>
               <p className='text-sm text-gray-400'>{timeRange}</p>
@@ -85,18 +74,13 @@ const OverView = () => {
       {/* Right Column (Desktop Only) */} 
       <div className="sticky z-10 self-start hidden lg:block top-20 h-fit">
         <section className="p-4 mb-4 bg-white border border-gray-300 dark:bg-dark dark:border-medium-dark dark:lg:bg-medium-dark rounded-xl ">
-          <OverViewPage />
+          <ExpensesTotal />
         </section>
 
         <div className="p-4 bg-white border border-gray-300 dark:bg-dark dark:border-medium-dark dark:lg:bg-medium-dark rounded-xl">
           <SpentSummary />
         </div>
       </div>
-
-      {/* Add Item Modal
-      <CustomModal>
-        <AddItem />
-      </CustomModal> */}
     </div>
   );
 };
